# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão de empresas, pesquisas, dashboards, auditoria e submissões anônimas.\n\nFluxo sugerido para clientes frontend:\n1) Autenticar em /auth/login\n2) Enviar token JWT no header Authorization: Bearer \u003ctoken\u003e\n3) Consumir endpoints protegidos em /api/v1\n",
        "title": "API de Pesquisa de Clima Organizacional",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Equipe Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/empresas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista empresas com suporte a limit e offset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empresas"
                ],
                "summary": "Listar empresas",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limite de itens",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Deslocamento",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova empresa no sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empresas"
                ],
                "summary": "Criar empresa",
                "parameters": [
                    {
                        "description": "Dados da empresa",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto.EmpresaCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/empresas/cnpj/{cnpj}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empresas"
                ],
                "summary": "Buscar empresa por CNPJ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CNPJ",
                        "name": "cnpj",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/empresas/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empresas"
                ],
                "summary": "Buscar empresa por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empresas"
                ],
                "summary": "Atualizar empresa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto.EmpresaUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "empresas"
                ],
                "summary": "Remover empresa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da empresa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "organizational-climate-survey_backend_internal_application_dto.EmpresaCreateRequest": {
            "type": "object",
            "required": [
                "cnpj",
                "nome_fantasia",
                "razao_social"
            ],
            "properties": {
                "cnpj": {
                    "description": "CNPJ formatado (obrigatório e com validação de tamanho)",
                    "type": "string",
                    "example": "12.345.678/0001-90"
                },
                "nome_fantasia": {
                    "description": "Nome comercial da empresa (obrigatório)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Acme"
                },
                "razao_social": {
                    "description": "Razão social registrada (obrigatória)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Acme Industria LTDA"
                }
            }
        },
        "organizational-climate-survey_backend_internal_application_dto.EmpresaUpdateRequest": {
            "type": "object",
            "properties": {
                "nome_fantasia": {
                    "description": "Novo nome fantasia (opcional)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Acme Brasil"
                },
                "razao_social": {
                    "description": "Nova razão social (opcional)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2,
                    "example": "Acme Brasil LTDA"
                }
            }
        },
...
        "organizational-climate-survey_backend_internal_application_dto_response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Codigo de erro para consumo de clientes",
                    "type": "string",
                    "example": "BAD_REQUEST"
                },
                "data": {
                    "description": "Dados retornados, genéricos"
                },
                "error": {
                    "description": "Mensagem de erro, se houver",
                    "type": "string",
                    "example": "Parâmetro inválido"
                },
                "message": {
                    "description": "Mensagem opcional para o usuário",
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "description": "Indica se a operação foi bem-sucedida",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "organizational-climate-survey_backend_internal_application_dto_response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": ""
                },
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "Consulta realizada com sucesso"
                },
                "pagination": {
                    "description": "Informações de paginação",
                    "allOf": [
                        {
                            "$ref": "#/definitions/organizational-climate-survey_backend_internal_application_dto_response.PaginationInfo"
                        }
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "organizational-climate-survey_backend_internal_application_dto_response.PaginationInfo": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Itens por página",
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "description": "Página atual",
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "description": "Total de itens",
                    "type": "integer",
                    "example": 120
                },
                "total_pages": {
                    "description": "Total de páginas",
                    "type": "integer",
                    "example": 6
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}